requirements.txt
fastapi
uvicorn
pydantic_settings
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
python-dotenv
jupyter
pandas
python-decouple
jinja2
requests
sqlmodel
alembic


